+++ assets/images.js
@@ -1,0 +1,14 @@
+import background from './src/background.svg'
+import logo from './src/logo.svg'
+import logo2 from './src/logo2.svg'
+
+const images = {
+	adaptive_icon: require('./src/adaptive-icon.png'),
+	favicon: require('./src/favicon.png'),
+	icon: require('./src/icon.png'),
+	splash: require('./src/splash.png'),
+	logo2: logo2,
+	logo: logo,
+	background: background,
+};
+export default images;
