+++ src/screens/LogInScreen.js
@@ -1,7 +1,7 @@
 import React from 'react';
-import { StyleSheet, SafeAreaView, TouchableOpacity } from 'react-native';
-import { NativeBaseProvider, Box, Text, Input, Icon, Button, HStack, Divider } from 'native-base';
-import Logo from '../logo2.svg';
+import { StyleSheet, TouchableOpacity } from 'react-native';
+import {  Box, Text, Input, Icon, Button, HStack, Divider } from 'native-base';
+import Images from '../../assets/images';
 import { Ionicons, AntDesign, FontAwesome5 } from '@expo/vector-icons';
 import { useNavigation } from '@react-navigation/native';
 
@@ -9,11 +9,9 @@ const LogInScreen = () => {
     const navigation = useNavigation();
 
     return (
-        <SafeAreaView style={styles.container}>
-            <NativeBaseProvider>
                 <Box style={styles.innerContainer}>
                     <Box style={styles.box1}>
-                        <Logo />
+                        <Images.logo2 />
                         <Text style={styles.text1}>Log In Now!</Text>
                         <Text style={styles.text2}>Lorem Ipsum is simply dummy text of the printing
                             and typesetting industry.
@@ -116,8 +114,6 @@ const LogInScreen = () => {
                         </HStack>
                     </Box>
                 </Box>
-            </NativeBaseProvider>
-        </SafeAreaView>
     );
 };
